+++ Debug/Debug/Src/main.cyclo
@@ -1,3 +1,7 @@
-../Src/main.c:12:6:finished_transmission	2
-../Src/main.c:22:6:message_received	1
-../Src/main.c:30:5:main	2
+../Inc/core_cm4.h:1688:22:__NVIC_EnableIRQ	2
+../Inc/core_cm4.h:1816:22:__NVIC_SetPriority	2
+../Src/main.c:34:6:USART1_IRQHandler	2
+../Src/main.c:53:6:enable_usart1_interrupt	1
+../Src/main.c:67:6:finished_transmission	2
+../Src/main.c:79:6:handle_received_char	1
+../Src/main.c:87:5:main	1
