+++ Debug/Debug/Src/main.d
@@ -1,8 +1,7 @@
 Src/main.o: ../Src/main.c ../Inc/stm32f303xc.h ../Inc/core_cm4.h \
  ../Inc/cmsis_version.h ../Inc/cmsis_compiler.h ../Inc/cmsis_gcc.h \
  ../Inc/mpu_armv7.h ../Inc/system_stm32f3xx.h ../Inc/serial.h \
- ../Inc/stm32f303xc.h ../Inc/uart_interrupt.h ../Inc/serial.h \
- ../Inc/system_stm32f3xx.h
+ ../Inc/stm32f303xc.h ../Inc/system_stm32f3xx.h
 ../Inc/stm32f303xc.h:
 ../Inc/core_cm4.h:
 ../Inc/cmsis_version.h:
@@ -12,6 +11,4 @@ Src/main.o: ../Src/main.c ../Inc/stm32f303xc.h ../Inc/core_cm4.h \
 ../Inc/system_stm32f3xx.h:
 ../Inc/serial.h:
 ../Inc/stm32f303xc.h:
-../Inc/uart_interrupt.h:
-../Inc/serial.h:
 ../Inc/system_stm32f3xx.h:
