+++ Debug/Debug/Src/main.su
@@ -1,3 +1,7 @@
-../Src/main.c:12:6:finished_transmission	24	static
-../Src/main.c:22:6:message_received	16	static
-../Src/main.c:30:5:main	40	static
+../Inc/core_cm4.h:1688:22:__NVIC_EnableIRQ	16	static
+../Inc/core_cm4.h:1816:22:__NVIC_SetPriority	16	static
+../Src/main.c:34:6:USART1_IRQHandler	16	static
+../Src/main.c:53:6:enable_usart1_interrupt	8	static,ignoring_inline_asm
+../Src/main.c:67:6:finished_transmission	24	static
+../Src/main.c:79:6:handle_received_char	16	static
+../Src/main.c:87:5:main	8	static
