+++ Debug/Debug/Src/subdir.mk
@@ -9,24 +9,21 @@ C_SRCS += \
 ../Src/main.c \
 ../Src/serial.c \
 ../Src/syscalls.c \
-../Src/sysmem.c \
-../Src/uart_interrupt.c 
+../Src/sysmem.c 
 
 OBJS += \
 ./Src/intergration.o \
 ./Src/main.o \
 ./Src/serial.o \
 ./Src/syscalls.o \
-./Src/sysmem.o \
-./Src/uart_interrupt.o 
+./Src/sysmem.o 
 
 C_DEPS += \
 ./Src/intergration.d \
 ./Src/main.d \
 ./Src/serial.d \
 ./Src/syscalls.d \
-./Src/sysmem.d \
-./Src/uart_interrupt.d 
+./Src/sysmem.d 
 
 
 # Each subdirectory must supply rules for building sources it contributes
@@ -36,7 +33,7 @@ Src/%.o Src/%.su Src/%.cyclo: ../Src/%.c Src/subdir.mk
 clean: clean-Src
 
 clean-Src:
-	-$(RM) ./Src/intergration.cyclo ./Src/intergration.d ./Src/intergration.o ./Src/intergration.su ./Src/main.cyclo ./Src/main.d ./Src/main.o ./Src/main.su ./Src/serial.cyclo ./Src/serial.d ./Src/serial.o ./Src/serial.su ./Src/syscalls.cyclo ./Src/syscalls.d ./Src/syscalls.o ./Src/syscalls.su ./Src/sysmem.cyclo ./Src/sysmem.d ./Src/sysmem.o ./Src/sysmem.su ./Src/uart_interrupt.cyclo ./Src/uart_interrupt.d ./Src/uart_interrupt.o ./Src/uart_interrupt.su
+	-$(RM) ./Src/intergration.cyclo ./Src/intergration.d ./Src/intergration.o ./Src/intergration.su ./Src/main.cyclo ./Src/main.d ./Src/main.o ./Src/main.su ./Src/serial.cyclo ./Src/serial.d ./Src/serial.o ./Src/serial.su ./Src/syscalls.cyclo ./Src/syscalls.d ./Src/syscalls.o ./Src/syscalls.su ./Src/sysmem.cyclo ./Src/sysmem.d ./Src/sysmem.o ./Src/sysmem.su
 
 .PHONY: clean-Src
